pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "byod-cli"
version = "1.0.0"
description = "Command-line interface for Lablytics BYOD - Secure biotech data processing with zero-knowledge encryption"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Lablytics", email = "support@lablytics.io" }
]
keywords = [
    "biotech",
    "encryption",
    "secure-computing",
    "nitro-enclave",
    "nextflow",
    "genomics",
    "proteomics",
    "zero-knowledge",
    "aws",
    "kms"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Security :: Cryptography",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.1.0",
    "cryptography>=41.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "boto3>=1.28.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "moto>=5.0.0",
    "responses>=0.24.0",
]

[project.scripts]
byod = "byod_cli.cli:main"

[project.urls]
Homepage = "https://lablytics.io"
Documentation = "https://docs.lablytics.io/cli"
Repository = "https://github.com/lablytics/byod-cli"
Issues = "https://github.com/lablytics/byod-cli/issues"
Changelog = "https://github.com/lablytics/byod-cli/releases"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "LICENSE",
    "README.md",
]

[tool.hatch.build.targets.wheel]
packages = ["src/byod_cli"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=byod_cli --cov-report=term-missing"
